requirements.txt
Django
djangorestframework
django-filter
drf-spectacular
drf-spectacular-sidecar
psycopg2-binary